requirements.txt
 # BlueDefenderX/requirements.txt
streamlit>=1.28.0
pyyaml>=6.0
requests>=2.31.0
pandas>=1.5.0
scikit-learn>=1.3.0
psutil>=5.9.0
# We'll add more as specific module needs arise (e.g., kafka, pandas, scikit-learn)
